+++ components/notifier/src/destinations/ms_teams.py
@@ -11,28 +11,20 @@ def build_notification_text(notification: Notification) -> str:
     """Create and format the contents of the notification."""
     nr_changed = len(notification.metrics)
     plural_s = "s" if nr_changed > 1 else ""
-    plural_are = "are" if nr_changed > 1 else "is"
     report_link = f"[{notification.report_title}]({notification.url})"
-    result = f"{report_link} has {nr_changed} metric{plural_s} that {plural_are} notable:\n\n"
-    for metric_notification_data in notification.metrics:
-        name = f"{metric_notification_data.subject_name}: {metric_notification_data.metric_name}"
-        new_status = metric_notification_data.new_metric_status
-        new_value = (
-            "?" if metric_notification_data.new_metric_value is None else metric_notification_data.new_metric_value
-        )
-        unit = metric_notification_data.metric_unit
-        unit = unit if unit.startswith("%") else f" {unit}"
-        if metric_notification_data.reason == "status_changed":
-            old_value = (
-                "?" if metric_notification_data.old_metric_value is None else metric_notification_data.old_metric_value
-            )
+    result = f"{report_link} has {nr_changed} metric{plural_s} that changed status:\n\n"
+    subjects = sorted(list(set(metric.subject_name for metric in notification.metrics)))
+    for subject in subjects:
+        result += f"* {subject}:\n"
+        for metric in [m for m in notification.metrics if m.subject_name == subject]:
+            new_value = "?" if metric.new_metric_value is None else metric.new_metric_value
+            old_value = "?" if metric.old_metric_value is None else metric.old_metric_value
+            unit = metric.metric_unit if metric.metric_unit.startswith("%") else f" {metric.metric_unit}"
+            old_value_text = " (unchanged)" if new_value == old_value else f", was {old_value}{unit}"
             result += (
-                f"* {name} status is "
-                f"{new_status}, was {metric_notification_data.old_metric_status}. "
-                f"Value is {new_value}{unit}, was {old_value}{unit}.\n"
+                f"  * *{metric.metric_name}* status is {metric.new_metric_status}, was {metric.old_metric_status}. "
+                f"Value is {new_value}{unit}{old_value_text}.\n"
             )
-        else:
-            result += f"* {name} has been {new_status} for three weeks. Value: {new_value}{unit}.\n"
     return result